Add API call for game schedule
Decided to use a separate API crate to keep things separated. The error handling and testing around the API are non-existent right now.

Longer term I would like to set up some automated tests that run on a schedule as a way to check that API client is working properly. Since there is no documentation for the API, I think this is the best way to ensure it will work.