Publish mlbt image to ghcr
This commit includes actions to publish images to the github
container registry on both pushes to the main branch and
tagged releases.

The labels, most importantly the org.opencontainers.image.source
is a requirement for public access to the built image.

NOTE: For users to be able to pull the image, the mlbt package
will need to be made public. This is located in package settings.

Signed-off-by: Isaac Boyd <isaacnboyd@gmail.com>